requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==7.1.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
deprecation==2.1.0
diffusers==0.21.4
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
Flask==1.1.4
flask-cors==6.0.1
flaskwebgui==1.1.6
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gotrue==2.12.3
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.20.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.15.0
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
lama-cleaner==0.17.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.5
libclang==18.1.1
loguru==0.7.3
Markdown==3.8.2
MarkupSafe==2.0.1
marshmallow==3.26.1
matplotlib==3.10.3
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.24.4
oauthlib==3.3.1
openai==1.95.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.18
packaging==24.2
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
postgrest==1.1.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyrate-limiter==3.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
realtime==2.5.3
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.7.0
scipy==1.15.3
sec-edgar-downloader==5.0.3
segment-anything==1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.167
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==1.0.1
win32_setctime==1.2.0
wrapt==1.14.1
yacs==0.1.8
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
feedparser
-U google-generativeai
matplotlib seaborn numpy google-generativeai
